split up components and reorganize
