+++ components/initialValues.ts
@@ -1,0 +1,5 @@
+export const countryList = [
+  {label: 'Afghanistan', value: 'Afghanistan'},
+  {label: 'Albania', value: 'Albania'},
+  {label: 'Algeria', value: 'Algeria'},
+];
