+++ screens/Form/initialValues.ts
@@ -1,8 +1,4 @@
-export const countryList = [
-  {label: 'Afghanistan', value: 'Afghanistan'},
-  {label: 'Albania', value: 'Albania'},
-  {label: 'Algeria', value: 'Algeria'},
-];
+import {countryList} from '../../components/initialValues';
 
 export const initialValues = {
   ssn: '',
