+++ screens/Form/types.ts
@@ -1,1 +1,8 @@
 export interface IUserForm {}
+
+export interface IFormValues {
+  ssn: string;
+  phoneNumber: string;
+  email: string;
+  country: string;
+}
